Major refactor to storing of accounts
Originally, accounts are stored in its own XML file which resulted in unnecessary duplication of codes.

It is realised that the accounts can be stored using the existing storage classes in the same file (addressbook.xml) which would solve the issue of duplicated codes.

Let's refactor and integrate all storing of accounts into the existing classes, and remove all unnecessary files.