AddressBook: Add undo command support for accounts
The addressbook currently only commits whenever it is manipulated for a person object only.
This means that if the addressbook is cleared and followed by an `undo` command,
it will only restore the person nodes and not the account nodes.

This may cause confusion amongst users on why the accounts data are not showing up.

Let's update the relevant classes to address this issue.