Password: Implement (unused) BCrypt hashing
At the moment, password will not be hashed (i.e. still stored in raw text), but the main functions are already roughly implemented.
There are some problems faced when password gets hashed. Every time a hash is generated, it uses a salt which is random.
So hashing the same password twice will result in two vastly different hashes. This raises an issue during unit testing such as testing
whether two Account objects are the same always returning false, or the Parser not equating despite the input being the same

Let's comment it out for now and leave it to the next milestone.