Password: Hash password
Passwords are now properly hashed and stored in the storage file. To keep it simple, salt is currently
generated using the username rather than using Random/SecureRandom. This limitation can be done
away in the future when the authentication portion is more mature.

In addition, raw password length has been reduced to 6~20. Anything more, and matching a certain pattern
is assumed to not be a raw password (i.e. hashed).

Resolves: #54