MainWindow: Include default person panel
If this line is removed, systemtests will fail as it will fail to find the nodes after filling the window's inner parts. This means that when the application is launched, the person list will still be shown, but will be automatically switched to the menu list upon logging in. This can only be resolved in the later milestone when we conduct a source-wide refactoring, that is, updating the systemtests to test menu instead of person, so we can then safely remove this line.