AddressBookParser: Parse deregister command
Since we are only interested in removing the account based on username, let's assume that if the username exists that the account exists as well. This means that we can use an Account constructor that only takes in a Username, and if it exists, then we retrieve the entire account based on the Username alone. This way, the admin won't have to get the password from the staff.