ModelManager: Set to default data rather than blank
When the `clear` command is executed, reset the xml to the default state with the root account. That is, everything else will be removed except for the root account so that the user will be able to login again to re-setup the system. The same logic also applies when the .xml is not found. This ensures that the system will be accessible in any state with the root account.

Please ignore the Persons that are still being populated as it's pending to be removed in future commits.